ops refactor
