+++ experimental/torch_xla2/torch_xla2/tensor.py
@@ -1,53 +1,16 @@
-import functools
+import contextlib
 import jax
 from jax import dlpack as jaxdl
 import jax.numpy as jnp
 import numpy
 import torch
 import torch.func
-import torch._decomp.decompositions
-from torch_xla2 import ops_registry
 import torch.utils._python_dispatch as torch_dispatch
 import torch.utils._pytree as torch_pytree
 import torch.utils.dlpack as torchdl
-from torch_xla2.ops import jaten
-from torch._subclasses.fake_tensor import FakeTensorMode
 
-fake_mode = FakeTensorMode()
-
-
-class XLADispatchMode(torch_dispatch.TorchDispatchMode):
-
-  def __torch_dispatch__(self, fn, types, args=(), kwargs=None):
-    if fn in constructors:
-      args, kwargs = unwrap((args, kwargs))
-      res = constructors[fn](*args, **kwargs)
-      return wrap(res)
-
-    return fn(*args, **kwargs)
-
-
-def _aten_arange(start,
-                 end,
-                 *,
-                 dtype=None,
-                 layout=None,
-                 requires_grad=False,
-                 device=None,
-                 pin_memory=False):
-  return jnp.arange(start, end, 1)
-
-
-def _aten_scalar_tensor(val, **kwargs):
-  p = torch.ops.aten.scalar_tensor(val)
-  return wrap(t2j(p))
-
-
-constructors = {
-    torch.ops.aten.scalar_tensor.default: _aten_scalar_tensor,
-    torch.ops.aten.arange.default: functools.partial(_aten_arange, 0),
-    torch.ops.aten.arange.start: _aten_arange,
-}
+class OperatorNotFound(Exception):
+  pass
 
 
 def wrap(jaxarray):
@@ -61,7 +24,9 @@ def unwrap(torchtensors):
 def t2j(t):
   if isinstance(t, XLATensor2):
     return t._elem
+  is_bool = False
   if t.dtype == torch.bool:
+    is_bool = True
     t = t.to(torch.int8)
 
   if not t.is_contiguous():
@@ -82,7 +47,7 @@ def t2j(t):
     if t.dtype == torch.bfloat16:
       res = res.astype(jnp.bfloat16)
 
-  if t.dtype == torch.bool:
+  if is_bool:
     res = res.astype(jnp.bool_)
   return res
 
@@ -97,48 +62,41 @@ def j2t(x):
     res = res.to(torch.bool)
   return res
 
+TORCH_DTYPE_TO_JAX = {
+      torch.float16: jnp.dtype('float16'),
+      torch.bfloat16: jnp.dtype('bfloat16'),
+      torch.half: jnp.dtype('float16'),
+      torch.float32: jnp.dtype('float32'),
+      torch.double: jnp.dtype('double'),
+      torch.long: jnp.dtype('int64'),
+      torch.int32: jnp.dtype('int32'),
+      torch.int16: jnp.dtype('int16'),
+      torch.int8: jnp.dtype('int8'),
+      torch.uint8: jnp.dtype('uint8'),
+      torch.bool: jnp.dtype('bool_'),
+      torch.complex64: jnp.dtype('complex64'),
+      torch.complex128: jnp.dtype('complex128'),
+      None: None,
+}
+
+JAX_DTYPE_TO_TORCH = {value: key for key, value in TORCH_DTYPE_TO_JAX.items()}
 
 def t2j_dtype(dtype):
-  return {
-      torch.float16: jnp.float16,
-      torch.bfloat16: jnp.bfloat16,
-      torch.half: jnp.float16,
-      torch.float32: jnp.float32,
-      torch.double: jnp.double,
-      torch.long: jnp.int64,
-      torch.int32: jnp.int32,
-      torch.int16: jnp.int16,
-      torch.int8: jnp.int8,
-      torch.uint8: jnp.uint8,
-      torch.bool: jnp.bool_,
-      torch.complex64: jnp.complex64,
-      torch.complex128: jnp.complex128,
-  }.get(dtype)
+  if dtype not in TORCH_DTYPE_TO_JAX:
+    raise RuntimeError(f'Attempting to convert unknown type: {dtype} to torch type,')
+  return TORCH_DTYPE_TO_JAX[dtype]
 
 
 def j2t_dtype(dtype):
-  return {
-      jnp.float16: torch.float16,
-      jnp.bfloat16: torch.bfloat16,
-      jnp.double: torch.double,
-      jnp.float32: torch.float32,
-      jnp.float16: torch.half,
-      jnp.int64: torch.long,
-      jnp.int32: torch.int32,
-      jnp.int16: torch.int16,
-      jnp.bool_: torch.bool,
-      jnp.complex64: torch.complex64,
-  }.get(dtype)
-
-
-def move_to_device(t):
-  return XLATensor2(t2j(t))
+  if dtype not in JAX_DTYPE_TO_TORCH:
+    raise RuntimeError(f'Attempting to convert unknown type: {dtype} to torch type,')
+  return JAX_DTYPE_TO_TORCH[dtype]
 
 
 class XLATensor2(torch.Tensor):
 
   @staticmethod
-  def __new__(cls, elem):
+  def __new__(cls, elem, env):
     dtype = j2t_dtype(elem.dtype)
     shape = list(elem.shape)
     for i, s in enumerate(shape):
@@ -154,9 +112,10 @@ class XLATensor2(torch.Tensor):
         requires_grad=False,
     )
 
-  def __init__(self, elem: jax.Array):
+  def __init__(self, elem: jax.Array, env: 'Environment'):
     super().__init__()
     self._elem = elem
+    self._env = env
 
   def __str__(self):
     return "XLATensor2({} {})".format(str(type(self._elem)), str(self._elem))
@@ -178,7 +137,7 @@ class XLATensor2(torch.Tensor):
     new_shape = (
         self._elem.shape[:start_dim] + (-1,) + self._elem.shape[end_dim:])
     new_elem = jnp.reshape(self._elem, new_shape)
-    return XLATensor2(new_elem)
+    return XLATensor2(new_elem, self._env)
     # return torch.reshape(self, new_shape)
 
   def __setitem__(self, key, val):
@@ -193,32 +152,17 @@ class XLATensor2(torch.Tensor):
 
   @classmethod
   def __torch_dispatch__(cls, func, types, args=(), kwargs=None):
-    kwargs = kwargs or {}
-    with jax.named_scope(func.name()):
+    env = None
+    for arg in torch_pytree.arg_tree_leaves(*args, **kwargs):
+      if isinstance(arg, XLATensor2):
+        env = arg._env
+        break
 
-      if isinstance(func, torch._ops.OpOverloadPacket):
-        return func(*args, **kwargs)
-
-      if func.name() == 'aten::copy_':
-        x, y = args
-        x._elem = y._elem
-        return
-
-      if func.overloadpacket in jaten.all_ops:
-        return jaten.all_ops[func.overloadpacket](*args, **kwargs)
-
-      lowering = ops_registry.lowerings.lookup(func)
-
-      if lowering is None:
-        raise RuntimeError("No lowering found for", func.name())
-
-      with XLADispatchMode():
-        res = lowering(*args, **kwargs)
-      debug_accuracy(func, args, kwargs, res)
-      return res
+    with env.mode():
+      return func(*args, **(kwargs or {}))
 
   def detach(self):
-    return XLATensor2(jax.lax.stop_gradient(self.jax()))
+    return XLATensor2(jax.lax.stop_gradient(self.jax()), self._env)
 
   def numpy(self) -> numpy.ndarray:
     import numpy as np
@@ -230,6 +174,20 @@ class XLATensor2(torch.Tensor):
 
   def torch(self) -> torch.Tensor:
     return j2t(self.jax())
+
+  def to(self, *args, **kwargs):
+    if len(args) == 1:
+      if isinstance(args[0], torch.dtype):
+        return XLATensor2(self._elem.astype(t2j_dtype(args[0])), self._env)
+    if 'dtype' in kwargs:
+      dtype = kwargs['dtype']
+      return XLATensor2(self._elem.astype(t2j_dtype(dtype)), self._env)
+    return self
+
+  @property
+  def dtype(self):
+    return j2t_dtype(self._elem.dtype)
+
 
 
 # TODO: slice of slice should also be another slice
@@ -281,3 +239,155 @@ def debug_accuracy(func, args, kwargs, current_output):
       pdb.set_trace()
 
   return True
+
+
+class XLAFunctionMode(torch.overrides.TorchFunctionMode):
+  """Context manager that dispatches torch function calls to JAX."""
+
+  def __init__(self, env):
+     self.env = env
+
+  def __torch_function__(self,
+                         func,
+                         types,
+                         args=(),
+                         kwargs=None) -> torch.Tensor:
+    try:
+      return self.env.dispatch(func, types, args, kwargs)
+    except OperatorNotFound:
+      return func(*args, **(kwargs or {}))
+
+
+class XLADispatchMode(torch_dispatch.TorchDispatchMode):
+
+  def __init__(self, env):
+    self.env = env
+
+  def __torch_dispatch__(self, func, types, args=(), kwargs=None):
+    if isinstance(func, torch._ops.OpOverloadPacket):
+      with self:
+        return func(*args, **kwargs)
+    if func.namespace != 'aten':
+      return func(*args, **kwargs)
+    return self.env.dispatch(func, types, args, kwargs)
+
+def _name_of_func(func):
+  if hasattr(func, 'name'):
+    return func.name()
+  return func.__name__
+
+
+class Environment:
+    """This class holds a set of configurations and "globals" needed
+
+    for executing torch program using jax.
+    Things included so far:
+
+    op registry
+    PRNGKey
+    Configs
+
+    Also helper functions to manipulate those.
+    """
+
+    _prng_key: jax.random.PRNGKey
+
+
+    def __init__(self, random_seed):
+        self._prng_key = jax.random.PRNGKey(random_seed)
+        self._function_mode = XLAFunctionMode(self)
+        self._dispatch_mode = XLADispatchMode(self)
+
+        # name is torch callable
+        self._ops = {}
+        self.load_ops()
+
+    def load_ops(self):
+      from torch_xla2.ops import jaten, jtorch, ops_registry
+      self._ops.update(ops_registry.all_aten_ops)
+      self._ops.update(ops_registry.all_torch_functions)
+
+      decomps = torch._decomp.core_aten_decompositions() 
+      from torch_xla2.decompositions import EXTRA_DECOMP
+      decomps.update(EXTRA_DECOMP)
+      for k, v in decomps.items():
+        if k not in self._ops:
+          self._ops[k] = ops_registry.Operator(
+            k,
+            v,
+            is_jax_function=False,
+            is_user_defined=False,
+            needs_env=False
+          )
+
+    def get_and_rotate_prng_key(self):
+        self._prng_key, key = jax.random.split(self._prng_key)
+        return key
+
+    def dispatch(self, func, types, args, kwargs):
+      with jax.named_scope(_name_of_func(func)):
+        kwargs = kwargs or {}
+        op = self._ops.get(func)
+
+        if op is None and isinstance(func, torch._ops.OpOverloadPacket):
+          op = self._ops.get(func.default)
+
+        if op is None and isinstance(func, torch._ops.OpOverload):
+          op = self._ops.get(func.overloadpacket)
+
+        if op is None:
+          raise OperatorNotFound(
+            f'Operator with name {_name_of_func(func)} has no lowering')
+
+        if op.is_jax_function:
+          args, kwargs = self.t2j_iso((args, kwargs))
+
+        if op.needs_env:
+          kwargs['env'] = self
+
+        with self.mode():
+          res = op.func(*args, **kwargs)
+
+        if op.is_jax_function:
+          res = self.j2t_iso(res)
+        
+        #if self.config.debug_accuracy_for_each_op:
+        #  debug_accuracy(func, args, kwargs, res)
+        return res
+
+    @contextlib.contextmanager
+    def mode(self):
+      with self._dispatch_mode, self._function_mode:
+        yield
+
+    def _move_one_value(self, val):
+      if isinstance(val, torch.nn.Module):
+        state_dict = self.to_xla(val.state_dict())
+        val.load_state_dict(state_dict, assign=True)
+        return val
+      if isinstance(val, XLATensor2):
+        return val
+      if isinstance(val, torch.Tensor):
+        return XLATensor2(t2j(val), self)
+      return val
+
+    def to_xla(self, torchvalues):
+      # tensors are torch.Tensors (not XLATensor)
+      res = torch_pytree.tree_map(
+        self._move_one_value,
+          torchvalues)
+      return res
+
+    def t2j_iso(self, torchtensors):
+      return torch_pytree.tree_map_only(
+        XLATensor2, lambda x: x.jax(), torchtensors)
+
+    def j2t_iso(self, jaxarray):
+      return torch_pytree.tree_map_only(
+        jnp.ndarray, lambda x: XLATensor2(x, self), jaxarray)
+
+    def j2t_copy(self, args):
+      pass
+
+    def j2t_copy(self, args):
+      pass
